LayoutTests/fast/dom/Window/postMessage-clone-port-error-expected.txt

PASS successfullyParsed is true

TEST COMPLETE
PASS exName is "DataCloneError"
